requirements.txt
faker
pandas
pyarrow
pyyaml
openpyxl
fastavro